requirements.txt
google-cloud-bigquery>=3.11.0
torch>=2.1.0
transformers>=4.35.0
pytorch-tabular>=1.0.0
pytorch-frame>=0.2.0
duckdb>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
faiss-cpu>=1.7.4
networkx>=3.1
py-elasticsearch>=8.10.0
neo4j>=5.14.0
kafka-python>=2.0.2
snorkel>=0.9.9
openai>=1.3.0
sentence-transformers>=2.2.2
pytorch-lightning>=2.1.0
tqdm>=4.66.0
psutil>=5.9.5
msgpack>=1.0.5
pyarrow>=14.0.0